+++ cli/cmd/migrate.go
@@ -1,15 +1,15 @@
 package cmd
 
 import (
-	"context"
 	"fmt"
 	"strings"
-	"time"
 
-	"github.com/cloudquery/plugin-sdk/clients"
+	discovery "github.com/cloudquery/plugin-sdk/clients/discovery/v0"
+	"github.com/cloudquery/plugin-sdk/registry"
 	"github.com/cloudquery/plugin-sdk/specs"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
+	"golang.org/x/exp/slices"
 )
 
 const (
@@ -59,74 +59,42 @@ func migrate(cmd *cobra.Command, args []string) error {
 			}
 			destinationsSpecs = append(destinationsSpecs, *spec)
 		}
-		if err := migrateConnection(ctx, cqDir, *sourceSpec, destinationsSpecs); err != nil {
-			return fmt.Errorf("failed to migrate source %s: %w", sourceSpec.Name, err)
+		discoveryClient, err := discovery.NewClient(ctx, sourceSpec.Registry, registry.PluginTypeSource, sourceSpec.Path, sourceSpec.Version)
+		if err != nil {
+			return fmt.Errorf("failed to create discovery client for source %s: %w", sourceSpec.Name, err)
 		}
-	}
-
-	return nil
-}
-
-func migrateConnection(ctx context.Context, cqDir string, sourceSpec specs.Source, destinationsSpecs []specs.Destination) error {
-	destinationNames := make([]string, len(destinationsSpecs))
-	for i := range destinationsSpecs {
-		destinationNames[i] = destinationsSpecs[i].Name
-	}
-	migrateTime := time.Now().UTC()
-
-	log.Info().Str("source", sourceSpec.Name).Strs("destinations", destinationNames).Time("migrate_time", migrateTime).Msg("Start migration")
-	defer log.Info().Str("source", sourceSpec.Name).Strs("destinations", destinationNames).Time("migrate_time", migrateTime).Msg("End migration")
-
-	sourceClient, err := clients.NewSourceClient(ctx, sourceSpec.Registry, sourceSpec.Path, sourceSpec.Version,
-		clients.WithSourceLogger(log.Logger),
-		clients.WithSourceDirectory(cqDir),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to get source plugin client for %s: %w", sourceSpec.Name, err)
-	}
-	defer func() {
-		if err := sourceClient.Terminate(); err != nil {
-			log.Error().Err(err).Msg("Failed to terminate source client")
-			fmt.Println("failed to terminate source client:", err)
+		versions, err := discoveryClient.GetVersions(ctx)
+		if err != nil {
+			if discoveryErr := discoveryClient.Terminate(); err != nil {
+				log.Error().Err(discoveryErr).Msg("failed to terminate discovery client")
+				fmt.Println("failed to terminate discovery client:", discoveryErr)
+			}
+			if err := migrateConnectionV0(ctx, cqDir, *sourceSpec, destinationsSpecs); err != nil {
+				return fmt.Errorf("failed to migrate source %s: %w", sourceSpec.Name, err)
+			}
+			continue
 		}
-	}()
 
-	destClients, err := newDestinationClients(ctx, sourceSpec, destinationsSpecs, cqDir)
-	if err != nil {
-		return err
-	}
-	defer destClients.Close()
-
-	selectedTables, tablesForSpecSupported, err := getTablesForSpec(ctx, sourceClient, sourceSpec)
-	if err != nil {
-		return fmt.Errorf("failed to get tables for source %s: %w", sourceSpec.Name, err)
-	}
-	tableCount := len(selectedTables.FlattenTables())
-
-	// Print a count of the tables that will be migrated.
-	if tablesForSpecSupported {
-		word := "tables"
-		if tableCount == 1 {
-			word = "table"
+		if err := discoveryClient.Terminate(); err != nil {
+			return fmt.Errorf("failed to terminate discovery client: %w", err)
 		}
-		fmt.Printf("Source %s will migrate %d %s.\n", sourceSpec.Name, tableCount, word)
-	}
 
-	fmt.Println("Starting migration for:", sourceSpec.Name, "->", sourceSpec.Destinations)
-	log.Info().Str("source", sourceSpec.Name).Strs("destinations", sourceSpec.Destinations).Msg("Starting migration")
-
-	for i, destinationSpec := range destinationsSpecs {
-		if err := destClients[i].Migrate(ctx, selectedTables); err != nil {
-			return fmt.Errorf("failed to migrate source %s on destination %s : %w", sourceSpec.Name, destinationSpec.Name, err)
+		if slices.Index(versions, "v1") != -1 {
+			if err := migrateConnectionV1(ctx, cqDir, *sourceSpec, destinationsSpecs); err != nil {
+				return fmt.Errorf("failed to migrate source %s: %w", sourceSpec.Name, err)
+			}
+			continue
 		}
+
+		if slices.Index(versions, "v0") != -1 {
+			if err := migrateConnectionV0(ctx, cqDir, *sourceSpec, destinationsSpecs); err != nil {
+				return fmt.Errorf("failed to migrate source %s: %w", sourceSpec.Name, err)
+			}
+			continue
+		}
+
+		return fmt.Errorf("failed to migrate source %s, unknown versions %v", sourceSpec.Name, versions)
 	}
-	tt := time.Since(migrateTime)
-	fmt.Printf("Migration completed successfully.\n")
-	log.Info().Str("source", sourceSpec.Name).
-		Strs("destinations", sourceSpec.Destinations).
-		Int("num_tables", tableCount).
-		Float64("time_took", tt.Seconds()).
-		Msg("Migration completed successfully")
 
 	return nil
 }