+++ plugins/destination/gcs/client/client_test.go
@@ -19,10 +19,11 @@ func TestPluginCSV(t *testing.T) {
 			NoRotate: true,
 		},
 		destination.PluginTestSuiteTests{
-			SkipOverwrite:     true,
-			SkipDeleteStale:   true,
-			SkipSecondAppend:  true,
-			SkipMigrateAppend: true,
+			SkipOverwrite:        true,
+			SkipDeleteStale:      true,
+			SkipSecondAppend:     true,
+			SkipMigrateAppend:    true,
+			SkipMigrateOverwrite: true,
 		},
 	)
 }
@@ -38,10 +39,11 @@ func TestPluginJSON(t *testing.T) {
 			NoRotate: true,
 		},
 		destination.PluginTestSuiteTests{
-			SkipOverwrite:     true,
-			SkipDeleteStale:   true,
-			SkipSecondAppend:  true,
-			SkipMigrateAppend: true,
+			SkipOverwrite:        true,
+			SkipDeleteStale:      true,
+			SkipSecondAppend:     true,
+			SkipMigrateAppend:    true,
+			SkipMigrateOverwrite: true,
 		},
 	)
 }