+++ plugins/destination/mssql/client/queries/create_table_test.go
@@ -13,7 +13,7 @@ func TestCreateTable(t *testing.T) {
 		expected   = `CREATE TABLE [cq].[table_name] (
   [_cq_id] uniqueidentifier UNIQUE NOT NULL,
   [_cq_parent_id] uniqueidentifier,
-  [_cq_source_name] varchar(8000),
+  [_cq_source_name] nvarchar(4000),
   [_cq_sync_time] datetimeoffset,
   [extra_col] float NOT NULL
   CONSTRAINT [table_name_cqpk] PRIMARY KEY (