+++ plugins/source/aws/resources/services/cloudtrail/trails.go
@@ -24,6 +24,9 @@ func Trails() *schema.Table {
 				Name:     "region",
 				Type:     schema.TypeString,
 				Resolver: client.ResolveAWSRegion,
+				CreationOptions: schema.ColumnCreationOptions{
+					PrimaryKey: true,
+				},
 			},
 			{
 				Name:     "cloudwatch_logs_log_group_name",