+++ plugins/source/azure/policies/queries/eventhub/namespaces_without_logging.sql
@@ -1,11 +1,11 @@
 WITH
     settings_with_logs AS (
-        SELECT resource_uri, JSONB_ARRAY_ELEMENTS(logs) AS logs FROM azure_monitor_diagnostic_settings
+        SELECT resource_id, JSONB_ARRAY_ELEMENTS(properties -> 'logs') AS logs FROM azure_monitor_diagnostic_settings
     ),
     namespaces_with_logging_enabled AS (
         SELECT DISTINCT n._cq_id
   FROM azure_eventhub_namespaces n
-    LEFT JOIN settings_with_logs s ON n.id = s.resource_uri
+    LEFT JOIN settings_with_logs s ON n.id = s.resource_id
     WHERE (s.logs->>'enabled')::boolean IS TRUE
 )
 insert into azure_policy_results
@@ -21,3 +21,4 @@ SELECT
   end
 FROM azure_eventhub_namespaces a
      LEFT JOIN namespaces_with_logging_enabled e ON a._cq_id = e._cq_id
+